memory: use unchecked indexing for PROGRAM_ADDRESS
`heap` is never shrunk and construction guarantees `heap.len() >= 1`,
so accessing `heap[PROGRAM_ADDRESS]` is always correct and never
panics. This means it's safe to use unchecked indexing.

We cannot use unchecked indexing for the program counter however, since
this can be manipulated by the executing program.

This appears to net a 10% performance improvement.