machine: remove use of bitpack
It was hard for me to follow what the bitpacking code was doing, and all
we really need to do is read the 4 most significant bits.

I think the bitpack module is now unused, but I've left it here in case
it's intended to be used for something else.

We also remove the Opcode::Err variant and just use Option<Opcode> where
applicable.